+++ core/src/main/java/ru/tinkoff/acquiring/sdk/models/enums/ResponseStatus.kt
@@ -39,6 +39,7 @@ enum class ResponseStatus {
     REFUNDED,
     PARTIAL_REFUNDED,
     REJECTED,
+    DEADLINE_EXPIRED,
     UNKNOWN,
     LOOP_CHECKING,
     COMPLETED,