+++ ui/src/main/java/ru/tinkoff/acquiring/sdk/models/NspkRequest.kt
@@ -16,6 +16,7 @@
 
 package ru.tinkoff.acquiring.sdk.models
 
+import ru.tinkoff.acquiring.sdk.responses.NspkC2bResponse
 import ru.tinkoff.acquiring.sdk.utils.NspkClient
 import ru.tinkoff.acquiring.sdk.utils.Request
 import kotlin.coroutines.resume
@@ -25,7 +26,7 @@ import kotlin.coroutines.suspendCoroutine
 /**
  * @author Mariya Chernyadieva
  */
-internal class NspkRequest : Request<NspkResponse> {
+internal class NspkRequest : Request<NspkC2bResponse> {
 
     @Volatile
     private var disposed = false
@@ -38,7 +39,7 @@ internal class NspkRequest : Request<NspkResponse> {
         disposed = true
     }
 
-    override fun execute(onSuccess: (NspkResponse) -> Unit, onFailure: (Exception) -> Unit) {
+    override fun execute(onSuccess: (NspkC2bResponse) -> Unit, onFailure: (Exception) -> Unit) {
         val client = NspkClient()
         client.call(this, onSuccess, onFailure)
     }