+++ ui/src/main/java/ru/tinkoff/acquiring/sdk/ui/activities/PaymentActivity.kt
@@ -17,43 +17,41 @@
 package ru.tinkoff.acquiring.sdk.ui.activities
 
 import android.annotation.SuppressLint
+import android.content.ActivityNotFoundException
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
 import android.view.View
 import androidx.lifecycle.Observer
+import ru.tinkoff.acquiring.sdk.AcquiringSdk
 import ru.tinkoff.acquiring.sdk.R
 import ru.tinkoff.acquiring.sdk.exceptions.AcquiringSdkException
 import ru.tinkoff.acquiring.sdk.exceptions.NetworkException
-import ru.tinkoff.acquiring.sdk.models.AsdkState
+import ru.tinkoff.acquiring.sdk.exceptions.NspkOpenException
+import ru.tinkoff.acquiring.sdk.models.*
 import ru.tinkoff.acquiring.sdk.models.BrowseFpsBankScreenState
-import ru.tinkoff.acquiring.sdk.models.BrowseFpsBankState
-import ru.tinkoff.acquiring.sdk.models.DefaultState
 import ru.tinkoff.acquiring.sdk.models.ErrorButtonClickedEvent
 import ru.tinkoff.acquiring.sdk.models.ErrorScreenState
 import ru.tinkoff.acquiring.sdk.models.FinishWithErrorScreenState
 import ru.tinkoff.acquiring.sdk.models.FpsBankFormShowedScreenState
 import ru.tinkoff.acquiring.sdk.models.FpsScreenState
-import ru.tinkoff.acquiring.sdk.models.FpsState
 import ru.tinkoff.acquiring.sdk.models.LoadState
 import ru.tinkoff.acquiring.sdk.models.LoadedState
 import ru.tinkoff.acquiring.sdk.models.LoadingState
 import ru.tinkoff.acquiring.sdk.models.OpenTinkoffPayBankScreenState
-import ru.tinkoff.acquiring.sdk.models.OpenTinkoffPayBankState
 import ru.tinkoff.acquiring.sdk.models.PaymentScreenState
 import ru.tinkoff.acquiring.sdk.models.RejectedCardScreenState
-import ru.tinkoff.acquiring.sdk.models.RejectedState
 import ru.tinkoff.acquiring.sdk.models.Screen
 import ru.tinkoff.acquiring.sdk.models.ScreenState
 import ru.tinkoff.acquiring.sdk.models.SingleEvent
 import ru.tinkoff.acquiring.sdk.models.ThreeDsScreenState
 import ru.tinkoff.acquiring.sdk.models.options.screen.PaymentOptions
-import ru.tinkoff.acquiring.sdk.redesign.sbp.util.SbpHelper
+import ru.tinkoff.acquiring.sdk.responses.NspkC2bResponse
 import ru.tinkoff.acquiring.sdk.threeds.ThreeDsHelper
 import ru.tinkoff.acquiring.sdk.ui.customview.NotificationDialog
 import ru.tinkoff.acquiring.sdk.ui.fragments.PaymentFragment
 import ru.tinkoff.acquiring.sdk.viewmodel.PaymentViewModel
-import java.lang.IllegalStateException
+import kotlin.IllegalStateException
 
 /**
  * @author Mariya Chernyadieva
@@ -133,7 +131,7 @@ internal class PaymentActivity : TransparentActivity() {
                     finishWithCancel()
                 } else {
                     data?.getStringExtra(EXTRA_SBP_BANK_PACKAGE_NAME)?.let { packageName ->
-                        openSbpDeepLinkInBank(packageName)
+                        openSbpPackage(packageName, checkNotNull(data.getStringExtra(EXTRA_SBP_BANK_DEEPLINK)))
                     }
                 }
             }
@@ -180,7 +178,7 @@ internal class PaymentActivity : TransparentActivity() {
                         ThreeDsHelper.Launch(this@PaymentActivity,
                             THREE_DS_REQUEST_CODE, options, screen.data, screen.transaction)
                     } catch (e: Throwable) {
-                        finishWithError(e)
+                        finishWithError(e, screen.data.paymentId)
                     }
                 }
                 is BrowseFpsBankScreenState -> openBankChooser(screen.deepLink, screen.banks)
@@ -193,7 +191,7 @@ internal class PaymentActivity : TransparentActivity() {
 
     private fun handleScreenState(screenState: ScreenState) {
         when (screenState) {
-            is FinishWithErrorScreenState -> finishWithError(screenState.error)
+            is FinishWithErrorScreenState -> finishWithError(screenState.error, screenState.paymentId)
             is ErrorScreenState -> {
                 if (asdkState is FpsState || asdkState is BrowseFpsBankState || asdkState is OpenTinkoffPayBankState) {
                     finishWithError(AcquiringSdkException(NetworkException(screenState.message)))
@@ -211,24 +209,39 @@ internal class PaymentActivity : TransparentActivity() {
     }
 
     @SuppressLint("QueryPermissionsNeeded")
-    private fun openBankChooser(deepLink: String, banks: Set<Any?>?) {
+    private fun openBankChooser(deepLink: String, banks: List<NspkC2bResponse.NspkAppInfo>?) {
         if (!banks.isNullOrEmpty()) {
-            val supportedBanks = SbpHelper.getBankApps(packageManager, deepLink, banks)
+            val supportedBanks = getBankApps(deepLink, banks)
             val intent = BankChooseActivity.createIntent(this, options, supportedBanks, deepLink)
             startActivityForResult(intent, SBP_BANK_CHOOSE_REQUEST_CODE)
         } else {
-            val intent = Intent(Intent.ACTION_VIEW)
-            intent.data = Uri.parse(deepLink)
-            val chooserIntent = Intent.createChooser(intent, getString(R.string.acq_fps_chooser_title))
-            startActivityForResult(chooserIntent, SBP_BANK_REQUEST_CODE)
+            val nspkOpenException = NspkOpenException(
+                throwable = IllegalStateException("nspk date are null")
+            )
+            finishWithError(nspkOpenException)
         }
     }
 
-    private fun openSbpDeepLinkInBank(packageName: String) {
-        val payload = (paymentViewModel.screenChangeEventLiveData.value?.value as BrowseFpsBankScreenState).deepLink
+    @SuppressLint("QueryPermissionsNeeded")
+    private fun getBankApps(link: String, banks: List<NspkC2bResponse.NspkAppInfo>?): BankChooseInfo {
+        val sbpIntent = Intent(Intent.ACTION_VIEW)
+        val appAndLinks = buildMap {
+            banks?.forEach { appInfo ->
+                val deepLink = prepareNspkDeeplinkWithScheme(appInfo.schema, link)
+                sbpIntent.setDataAndNormalize(deepLink)
+                packageManager.queryIntentActivities(sbpIntent, 0).forEach {
+                    put(it.activityInfo.packageName, deepLink.toString())
+                }
+            }
+        }
+        return BankChooseInfo(appAndLinks)
+    }
+
+    private fun openSbpDeepLinkInBank(deepLink: String) {
         val intent = Intent(Intent.ACTION_VIEW)
-        intent.data = Uri.parse(payload)
-        intent.setPackage(packageName)
+        intent.data = Uri.parse(deepLink)
+        intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK
+        AcquiringSdk.log("try open intent with : Intent.ACTION_VIEW -d \"${intent.data}\"")
         startActivityForResult(intent, SBP_BANK_REQUEST_CODE)
     }
 
@@ -246,10 +259,39 @@ internal class PaymentActivity : TransparentActivity() {
         }
     }
 
+    private fun prepareNspkDeeplinkWithScheme(schema: String, deepLink: String): Uri {
+        val raw = Uri.parse(deepLink)
+        return Uri.Builder().apply {
+            this.scheme(schema)
+            this.authority(raw.authority)
+            this.path(raw.path)
+        }.build()
+    }
+
+    private fun getBrowseFpsBankScreenState(): BrowseFpsBankScreenState {
+       return paymentViewModel.screenChangeEventLiveData.value?.value as BrowseFpsBankScreenState
+    }
+
+    private fun openSbpPackage(packageName: String, deepLink: String) {
+        val fpsState = getBrowseFpsBankScreenState()
+        try {
+            openSbpDeepLinkInBank(deepLink)
+        } catch (e: Exception) {
+            val nspkOpenException = NspkOpenException(
+                throwable = e,
+                message = "$packageName cannot open via deeplink $deepLink",
+                deeplink = deepLink,
+                paymentId = fpsState.paymentId
+            )
+            finishWithError(nspkOpenException,  fpsState.paymentId)
+        }
+    }
+
     companion object {
         private const val SBP_BANK_REQUEST_CODE = 112
         private const val SBP_BANK_CHOOSE_REQUEST_CODE = 113
 
         internal const val EXTRA_SBP_BANK_PACKAGE_NAME = "sbp_bank_package_name"
+        internal const val EXTRA_SBP_BANK_DEEPLINK = "sbp_bank_deeplink"
     }
 }