+++ ui/src/main/java/ru/tinkoff/acquiring/sdk/viewmodel/ViewModelProviderFactory.kt
@@ -41,14 +41,9 @@ internal class ViewModelProviderFactory(
         AttachCardViewModel::class.java to AttachCardViewModel(application, handleErrorsInSdk, sdk),
         QrViewModel::class.java to QrViewModel(application, handleErrorsInSdk, sdk),
         ThreeDsViewModel::class.java to ThreeDsViewModel(application, handleErrorsInSdk, sdk),
-        NotificationPaymentViewModel::class.java to NotificationPaymentViewModel(
-            application,
-            handleErrorsInSdk,
-            sdk
-        ),
-        YandexPaymentViewModel::class.java to YandexPaymentViewModel(application, handleErrorsInSdk, sdk)
+        SavedCardsViewModel::class.java to SavedCardsViewModel(application, handleErrorsInSdk, sdk),
+        NotificationPaymentViewModel::class.java to NotificationPaymentViewModel(application, handleErrorsInSdk, sdk)
     )
-
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         return (viewModelCollection)[modelClass] as T