+++ yandexpay/src/main/java/ru/tinkoff/acquiring/yandexpay/YandexPayResult.kt
@@ -6,7 +6,7 @@ import ru.tinkoff.acquiring.sdk.models.options.screen.PaymentOptions
  * Created by i.golovachev
  */
 sealed class AcqYandexPayResult {
-    class Success(val token: String, val paymentOptions: PaymentOptions) : AcqYandexPayResult()
+    class Success(internal val token: String, val paymentOptions: PaymentOptions) : AcqYandexPayResult()
     class Error(val throwable: Throwable) : AcqYandexPayResult() {
 
         constructor(message: String) : this(YandexPayError(message))